refactor: move all Main components in one folder
