+++ src/components/UI/MainComponents/MainCard/MainCard.module.scss
