+++ src/components/UI/MainComponents/MainModal/MainModal.module.scss
